+++ packages/rx_bloc_cli/mason_templates/bricks/feature_onboarding/__brick__/bin/server/repositories/users_repository.dart
@@ -6,9 +6,13 @@ import 'package:{{project_name}}/base/models/confirmed_credentials_model.dart';
 import 'package:{{project_name}}/base/models/user_model.dart';
 import 'package:{{project_name}}/base/models/user_role.dart';
 
+const kPasswordResetTimeoutInSeconds = 60;
+
 class UsersRepository {
   final List<UserModel> _registeredUsers = [];
-final Map<String, String> _unconfirmedPhoneNumbers = {};
+  final Map<String, String> _passwords = {};
+  final Map<String, String> _unconfirmedPhoneNumbers = {};{{#enable_forgotten_password}}
+  final Map<String, int> _passwordResetLockedUsers = {};{{/enable_forgotten_password}}
 
   List<UserModel> getUsers() => _registeredUsers;
 
@@ -17,6 +21,8 @@ final Map<String, String> _unconfirmedPhoneNumbers = {};
 
   UserModel? getUserByEmail(String email) =>
       _registeredUsers.firstWhereOrNull((user) => user.email == email);
+
+  bool isUserRegistered(String email) => getUserByEmail(email) != null;
 
   void createUser(UserModel user) => _registeredUsers.add(user);
 
@@ -66,4 +72,24 @@ final Map<String, String> _unconfirmedPhoneNumbers = {};
 
   void deleteUser(String id, UserRole role) => _registeredUsers
       .removeWhere((user) => (user.id == id && user.role == role));
+
+  void setPasswordForUser(String email, String password) =>
+      _passwords[email] = password;
+
+  String? getPasswordForUser(String email) => _passwords[email];{{#enable_forgotten_password}}
+
+  bool isPasswordResetLockedForUser(String email) =>
+      _passwordResetLockedUsers.keys.contains(email);
+
+  int getPasswordResetTimeoutForUser(String email) =>
+      _passwordResetLockedUsers[email] ?? kPasswordResetTimeoutInSeconds;
+
+  void lockPasswordResetForUser(String email) =>
+      _passwordResetLockedUsers[email] = kPasswordResetTimeoutInSeconds;
+
+  void decrementPasswordResetTimeoutForUser(String email) =>
+      _passwordResetLockedUsers.update(email, (timeout) => --timeout);
+
+  void unlockPasswordResetForUser(String email) =>
+      _passwordResetLockedUsers.remove(email);{{/enable_forgotten_password}}
 }