+++ packages/rx_bloc_cli/mason_templates/bricks/feature_password_reset/__brick__/lib/base/models/request_models/password_reset_init_request_model.dart
@@ -1,0 +1,18 @@
+{{> licence.dart }}
+
+import 'package:json_annotation/json_annotation.dart';
+
+part 'password_reset_init_request_model.g.dart';
+
+@JsonSerializable()
+class PasswordResetInitRequestModel {
+  PasswordResetInitRequestModel(this.email);
+
+  /// The email of the user requesting the password reset
+  final String email;
+
+  factory PasswordResetInitRequestModel.fromJson(Map<String, dynamic> json) =>
+      _$PasswordResetInitRequestModelFromJson(json);
+
+  Map<String, dynamic> toJson() => _$PasswordResetInitRequestModelToJson(this);
+}