+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/bin/server/di/server_dependencies.dart
@@ -5,7 +5,8 @@ import '../controllers/authentication_controller.dart';{{/has_authentication}}{{
 import '../controllers/count_controller.dart';{{/enable_feature_counter}}{{#enable_feature_onboarding}}
 import '../controllers/country_codes_controller.dart';{{/enable_feature_onboarding}}{{#enable_feature_deeplinks}}
 import '../controllers/deep_links_controller.dart';{{/enable_feature_deeplinks}}{{#enable_mfa}}
-import '../controllers/mfa_controller.dart';{{/enable_mfa}}
+import '../controllers/mfa_controller.dart';{{/enable_mfa}}{{#enable_forgotten_password}}
+import '../controllers/password_reset_controller.dart';{{/enable_forgotten_password}}
 import '../controllers/permissions_controller.dart';{{#enable_pin_code}}
 import '../controllers/pin_code_controller.dart';{{/enable_pin_code}}
 import '../controllers/push_notifications_controller.dart';{{#enable_feature_onboarding}}
@@ -56,7 +57,7 @@ class ServerDependencies{
     {{#enable_feature_counter}}
     ..addController(CountController())
     {{/enable_feature_counter}}{{#has_authentication}}
-    ..addController(AuthenticationController(di.get())){{/has_authentication}}
+    ..addController(AuthenticationController(di.get(){{#enable_feature_onboarding}}, di.get(){{/enable_feature_onboarding}})){{/has_authentication}}
     ..addController(PushNotificationsController())
     ..addController(PermissionsController({{#has_authentication}}di.get(),{{#enable_feature_onboarding}} di.get(){{/enable_feature_onboarding}}{{/has_authentication}}))
     {{#enable_feature_deeplinks}}
@@ -69,7 +70,8 @@ class ServerDependencies{
     ..addController(UsersController(di.get(), di.get()))
     ..addController(RegistrationController(di.get(), di.get()))
     ..addController(CountryCodesController(di.get()))
-    {{/enable_feature_onboarding}}
+    {{/enable_feature_onboarding}}{{#enable_forgotten_password}}
+    ..addController(PasswordResetController(di.get()));{{/enable_forgotten_password}}
     ;
 
     /// TODO: Add your controllers here